Fix: iir filter returning uint16

+++ Core/Src/iirfilter.c
@@ -33,7 +33,7 @@ void iir_filter(uint16_t * input, uint16_t * output, uint8_t channel)
     y_state[channel] = ((c[coeff_index] * (int64_t)*input ) + ((100000000LL-c[coeff_index]) * y_state[channel]))/ 100000000LL;
 
     // the factor 1000000 is used 
-    *output = (int32_t)(y_state[channel]);
+    *output = (uint16_t)(y_state[channel]);
 }
 
 void iir_reset()
